build for ubuntu12.04 gcc4.6 merge from  harmy:master[https://github.com/kbengine/kbengine/pull/28]

+++ kbe/src/lib/entitydef/datatype.cpp
@@ -413,7 +413,7 @@ bool VectorType::isSameType(PyObject* pyValue)
 		return false;
 	}
 
-	if(!PySequence_Check(pyValue) || PySequence_Size(pyValue) != elemCount_)
+	if(!PySequence_Check(pyValue) || (uint32)PySequence_Size(pyValue) != elemCount_)
 	{
 		PyErr_Format(PyExc_TypeError, 
 			"must be set to a VECTOR%d type.", elemCount_);
